pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyrosense"
version = "0.1.0"
description = "Wildfire prediction using Prithvi foundation model and multi-source feature fusion"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "PyroSense Team" }
]
keywords = [
    "wildfire",
    "fire-prediction",
    "remote-sensing",
    "foundation-model",
    "prithvi",
    "earth-observation",
    "machine-learning",
    "stacking-ensemble",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "xarray>=2023.1.0",
    "zarr>=2.14",
    "rasterio>=1.3",
    "earthaccess>=0.5",
    "timm>=0.9",
    "loguru>=0.7",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
earth-engine = [
    "earthengine-api>=0.1.0",
]
earthdial = [
    "transformers>=4.37.0,<5.0.0",  # EarthDial requires <5.0
    "accelerate>=0.25.0",
    "peft>=0.8.0,<0.10.0",  # Compatible with transformers 4.37.x
    "sentencepiece",
    "einops",
    "pillow>=10.0",
]
dev = [
    "pytest>=7.3",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = [
    "mkdocs>=1.4",
    "mkdocs-material>=9.0",
]
all = [
    "pyrosense[earth-engine,earthdial,dev,docs]",
]

[project.scripts]
pyrosense = "pyrosense.cli.main:cli"

[project.urls]
Homepage = "https://github.com/georgepap23/pyrosense"
Documentation = "https://pyrosense.readthedocs.io"
Repository = "https://github.com/georgepap23/pyrosense"
Issues = "https://github.com/georgepap23/pyrosense/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pyrosense"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/configs",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # `zip()` without an explicit `strict=` parameter
]

[tool.ruff.lint.isort]
known-first-party = ["pyrosense"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "venv",
    "build",
    "dist",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]

[tool.coverage.run]
source = ["src/pyrosense"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
